\name{mixeddiffusion-internal}
\title{Internal functions}
\description{
Internal helper functions used by the EM and MCMC algorithms.
These functions are not intended to be called directly by users.
}
\alias{Mu_cond_at1}
\alias{Mu_cond_at2}
\alias{Mu_cond_const}
\alias{Mu_cond_cos}
\alias{Mu_cond_exp}
\alias{Mu_cond_sin}
\alias{Mu_dlt_cos}
\alias{Mu_dlt_exp}
\alias{Mu_dlt_power}
\alias{Mu_dlt_sin}
\alias{Q_lambda}
\alias{S2cond}
\alias{extract_p_power}
\alias{make_Mu_cond}
\alias{make_Mu_dlt}
\alias{posterior_ou}
\alias{posterior_winner}
\alias{update_sigma2}
\keyword{internal}
\value{No return value, called for side effects}
